requirements.txt
transformers
torch
llama-cpp-python
streamlit
sentencepiece
huggingface_hub